Refactor: Rename PatchWindow to AppWindow
