+++ mfp/gui_command.py
@@ -59,14 +59,14 @@ class GUICommand:
 
     def create(self, obj_type, obj_args, obj_id, parent_id, params):
         from .gui_main import MFPGUI
-        from .gui.patch_element import PatchElement
+        from .gui.base_element import BaseElement
         from .gui.processor_element import ProcessorElement
         from .gui.message_element import MessageElement
         from .gui.text_element import TextElement
         from .gui.enum_element import EnumElement
         from .gui.plot_element import PlotElement
         from .gui.slidemeter_element import SlideMeterElement, DialElement
-        from .gui.patch_info import PatchInfo
+        from .gui.patch_display import PatchDisplay
         from .gui.via_element import SendViaElement, ReceiveViaElement
         from .gui.via_element import SendSignalViaElement, ReceiveSignalViaElement
         from .gui.button_element import ToggleButtonElement
@@ -84,7 +84,7 @@ class GUICommand:
             'plot': PlotElement,
             'slidemeter': SlideMeterElement,
             'dial': DialElement,
-            'patch': PatchInfo,
+            'patch': PatchDisplay,
             'sendvia': SendViaElement,
             'recvvia': ReceiveViaElement,
             'sendsignalvia': SendSignalViaElement,
@@ -100,12 +100,12 @@ class GUICommand:
             o.parent_id = parent_id
             o.obj_type = obj_type
             o.obj_args = obj_args
-            o.obj_state = PatchElement.OBJ_COMPLETE
+            o.obj_state = BaseElement.OBJ_COMPLETE
 
-            if isinstance(o, PatchElement):
+            if isinstance(o, BaseElement):
                 parent = MFPGUI().recall(o.parent_id)
                 layer = None
-                if isinstance(parent, PatchInfo):
+                if isinstance(parent, PatchDisplay):
                     if "layername" in params:
                         layer = parent.find_layer(params["layername"])
                     if not layer:
@@ -113,7 +113,7 @@ class GUICommand:
                     layer.add(o)
                     # layer.backend.group.add_actor(o)
                     # o.container = layer.backend.group
-                elif isinstance(parent, PatchElement):
+                elif isinstance(parent, BaseElement):
                     # FIXME: don't hardcode GOP offsets
                     if not parent.export_x:
                         log.debug(
@@ -139,7 +139,7 @@ class GUICommand:
     def connect(self, obj_1_id, obj_1_port, obj_2_id, obj_2_port):
         from .gui_main import MFPGUI
         from .gui.connection_element import ConnectionElement
-        from .gui.patch_info import PatchInfo
+        from .gui.patch_display import PatchDisplay
         from mfp import log
 
         obj_1 = MFPGUI().recall(obj_1_id)
@@ -149,8 +149,8 @@ class GUICommand:
             log.debug("ERROR: connect: obj_1 (id=%s) --> %s, obj_2 (id=%s) --> %s"
                       % (obj_1_id, obj_1, obj_2_id, obj_2))
             return None
-        elif isinstance(obj_1, PatchInfo) or isinstance(obj_2, PatchInfo):
-            log.debug("Trying to connect a PatchInfo (%s [%s] --> %s [%s])"
+        elif isinstance(obj_1, PatchDisplay) or isinstance(obj_2, PatchDisplay):
+            log.debug("Trying to connect a PatchDisplay (%s [%s] --> %s [%s])"
                       % (obj_1.obj_name, obj_1_id, obj_2.obj_name, obj_2_id))
             return None
 
@@ -165,9 +165,9 @@ class GUICommand:
 
     async def delete(self, obj_id):
         from .gui_main import MFPGUI
-        from .gui.patch_info import PatchInfo
+        from .gui.patch_display import PatchDisplay
         obj = MFPGUI().recall(obj_id)
-        if isinstance(obj, PatchInfo):
+        if isinstance(obj, PatchDisplay):
             await obj.delete()
             if obj in MFPGUI().appwin.patches:
                 MFPGUI().appwin.patches.remove(obj)
@@ -176,9 +176,9 @@ class GUICommand:
 
     async def select(self, obj_id):
         from .gui_main import MFPGUI
-        from .gui.patch_info import PatchInfo
+        from .gui.patch_display import PatchDisplay
         obj = MFPGUI().recall(obj_id)
-        if isinstance(obj, PatchInfo):
+        if isinstance(obj, PatchDisplay):
             MFPGUI().appwin.layer_select(obj.layers[0])
         else:
             await MFPGUI().appwin.select(obj)
